new changes on env

+++ index.js
@@ -1,7 +1,11 @@
-//  abrir o projeto inicia o projeto = npm init |procsso que automatiza npm init  - y//
+// abrir o projeto inicia o projeto = npm init |procsso que automatiza npm init  - y//
 //npm i instalar
 //npm i express
 //npm i  nodemon -D ele instala so nas dev dependecs
+
+// baixaro projeto front abrir no cmd
+//abrir terminal npm i
+//npm i e audit fix por yes
 
 require("dotenv").config();
 const express = require("express");
@@ -9,16 +13,18 @@ const mongodb = require("mongodb");
 const ObjectId = mongodb.ObjectId;
 
 (async () => {
-	const dbhost= process.env.DB_HOST;
-	const dbport= process.env.DB_PORT;
-	const dbname=process.env.DB_NAME;
+	const dbUser= process.env.DB_USER;
+	const dbPassword= process.env.DB_PASSWORD;
+	const dbName=process.env.DB_NAME;
+	const dbChar=process.env.DB_CHAR;
 
 	const app = express();
 
 	app.use(express.json());
+
 	const port = process.env.PORT|| 3000;
 
-	const connectionString = `mongodb://${dbhost}:${dbport}/${dbname}`;
+	const connectionString = `mongodb+srv://${dbUser}:${dbPassword}@cluster0.${dbChar}.mongodb.net/${dbName}?retryWrites=true&w=majority`;
 
 	const options = {
 		useUnifiedTopology: true,
@@ -34,6 +40,8 @@ const ObjectId = mongodb.ObjectId;
 	const getPersonagemById = async (id) =>
 
 		personagens.findOne({ _id: ObjectId(id) });
+
+		//CORS É MT IMPORTANTE REVISE BIA!//
 
 		app.all("/*", (req, res, next) => {
 			res.header("Access-Control-Allow-Origin", "*");
@@ -68,12 +76,13 @@ app.post("/personagens", async(req, res)=> {
 	const objeto= req.body;
 
 	if(!objeto || !objeto.nome|| !objeto.imagemUrl){
-		res.send("Objeto invalid");
+		res.send("a requisicao nao esta valida, verifique se há campo de nome e imagemUrl");
 		return;
 	}
 	const insertCount= await personagens.insertOne(objeto);
-	console.log(insertCount);
-	if(!insertCount){
+
+	console.log(result);
+	if(result.acknowledge == false){
 		res.send("Ocorreu um erro");
 		return;
 	}
@@ -83,27 +92,60 @@ app.post("/personagens", async(req, res)=> {
 app.put("/personagens/:id", async (req, res) => {
 	const id= req. params.id;
 	const objeto=req.body;
-	res.send(
-		await personagens.updateOne(
+
+	if(!objeto || !objeto.nome || !objeto.imagemUrl){
+		res.send(
+			"a requisicao nao esta valida, verifique se há campo de nome e imagemUrl"
+		);
+		return;
+	}
+	const quantidadePersonagens= await personagens.countDocuments({
+		_id:ObjectId(id),
+	});
+
+	if (quantidadePersonagens !==1){
+		res.send("Personagem nao encontrado");
+		return;
+	}
+
+		 const result = await personagens.updateOne(
 			{
 				_id:ObjectId(id),
 			},
 			{
 				$set:objeto,
 			}
-		)
+		
 	);
 
+	if (result.modifiedCount !==1){
+		res.send("Ocorreu um erro ao tentar atualizar o personagem");
+		return;
+	}
+res.send(await getPersonagemById (id));
 });
 
 app.delete("/personagens/:id", async (req, res) => {
 	const id = req.params.id;
+const quantidadePersonagens= await personagens.countDocuments({
+	_id:Object(id),
 
-	res.send(
-		await personagens.deleteOne({
+});
+if (quantidadePersonagens !==1) {
+	res.send("personagem nao foi encontrado");
+	return;
+}
+	
+		 const result= await personagens.deleteOne({
 			_id:ObjectId(id),
-		})
-	);
+		});
+	// caso ocorra um erro e o personagem nao seja removido///
+
+	if (result.deletedCount !==1){
+		res.send("Ocorreu um erro ao tentar deletar o personagem");
+		return;
+	}
+	res.send("Pesonagem removido com sucsso");
 
 });
